entities package renamed in models, entities configuration & relations (wip)
