+++ src/main/java/fr/cogip/cybercogip/data/CustomerRepository.java
@@ -1,6 +1,6 @@
 package fr.cogip.cybercogip.data;
 
-import fr.cogip.cybercogip.entities.Customer;
+import fr.cogip.cybercogip.models.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository extends JpaRepository<Customer, Long> {
