--- src/main/java/fr/cogip/cybercogip/entities/PruductStatus.java
@@ -1,4 +1,0 @@
-package fr.cogip.cybercogip.entities;
-
-public class PruductStatus {
-}
