+++ src/main/java/fr/cogip/cybercogip/models/PruductStatus.java
@@ -1,0 +1,4 @@
+package fr.cogip.cybercogip.models;
+
+public class PruductStatus {
+}
